refactor: extract menu operations module
